api/requirements.txt
fastapi
uvicorn
SQLAlchemy
cryptography
python-dotenv
python-jose
loguru
passlib
alembic
asyncpg
slowapi
prometheus-fastapi-instrumentator
prometheus_client
pydantic
httpx
python-dateutil
pydantic[email]
psycopg2-binary

